# Use FileUpdateChecker for Migration::CheckPending 