This is a second attempt at #29759, which I should have got around to sooner, it wasn't that hard to fix 😅.

`Migration::CheckPending` is a rack middleware normally used in development to raise an exception when there are pending migrations.

This commit replaces the existing caching, which avoided checking all migrations if the newest migration hasn't changed. Instead, we now use `FileUpdateChecker`, which has two advantages: it can detect new migrations which aren't the highest version number, and it is faster.


## Improved checking

The existing code watched for the value of `last_migration.mtime` to change, the modification time on the  file with the highest version number. This works most of the time, but when merging/rebasing/checking out other branches, it's likely to have "new" migrations arrive with a lower version number than the previous highest. In this event the `CheckPending` will not notice and not re-check that migrations are up to date.

What we really want is to check for the greatest `mtime` of all migrations, and re-check if any files are new. This happens to be exactly what `ActiveSupport::FileUpdateChecker` already does! :tada: 

## Improved performance

This middleware is probably only run in development, but this check was actually quite slow (not shown in logs because the middleware came first). As an example I tested Discourse (900+ migrations) which added 25ms extra per request in development mode.

This slowness is because `last_migration` is `migrations.last`, and `migrations` needs to glob and then for each migration file needs to `File.basename`, a regex to extract version/name/scope, and `camelize`. Using `FileUpdateChecker` is a bit faster in my testing.

In the base case, we'll use `EventedFileUpdateChecker` (default for new apps), 

## Performance testing

For testing, I made a new rails application, and generated 500 migrations

```
$ bundle exec rails new --dev sandbox
$ bundle exec rails c
> require "rails/generators"
> require "rails/generators/rails/migration/migration_generator"
> 500.times { |i| Rails::Generators::MigrationGenerator.start(["TestSomeMigrations#{i}", "--orm=active_record"]) }
$ bundle exec rake db:migrate
```

I have the following benchmark script
``` ruby
require "benchmark/ips"
require "./config/environment"

check_pending = ActiveRecord::Migration::CheckPending.new(->(env){})
# check_pending = ActiveRecord::Migration::CheckPending.new(->(env){}, file_watcher: ActiveSupport::EventedFileUpdateChecker)

env = {}

Benchmark.ips do |x|
  x.config(:time => 5, :warmup => 2)

  x.report("CheckPending#call") { check_pending.call(env) }
end
```

**Original**

```
$ be ruby benchmark.rb
Warming up --------------------------------------
   CheckPending#call    11.000  i/100ms
Calculating -------------------------------------
   CheckPending#call    135.510  (±12.5%) i/s -    671.000  in   5.083931s
```

**This PR** (when `file_watcher == FileUpdateChecker`)

```
$ be ruby benchmark.rb
Warming up --------------------------------------
   CheckPending#call    34.000  i/100ms
Calculating -------------------------------------
   CheckPending#call    348.063  (±18.1%) i/s -      1.666k in   5.024624s
```

**This PR** (when `file_watcher == EventedFileUpdateChecker`)

(It basically becomes a no-op, because the file changes are reported async)

```
$ be ruby benchmark.rb
Warming up --------------------------------------
   CheckPending#call   128.936k i/100ms
Calculating -------------------------------------
   CheckPending#call      1.884M (± 2.5%) i/s -      9.412M in   5.000561s
```